requirements.txt
python-dateutil>=2.6.0
Markdown>=2.6.7
mutagen>=1.35.1
plumbum>=1.6.2
PyYAML>=3.12
unidecode>=0.4.19
